sale_deed_processor/backend_v1_backup/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.109.0
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
greenlet==3.2.4
groq==0.36.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==1.26.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
protobuf==6.33.1
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pypdfium2==5.1.0
pyreadline3==3.5.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
sympy==1.14.0
torch==2.1.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.27.0
watchfiles==1.1.1
websockets==15.0.1
